+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
+## [unreleased]
+### Added
+- Add an option to rhocall annotate to select 
+
 ## [0.5.2]
 - HW and AZ are flags, set Number appropriately in VCF
 - Use a pyproject.toml, hatchling build